Moved lang outside of GCT project

+++ examples/lang/build.ps1
@@ -3,7 +3,9 @@ Write-Host "Compiling stdio lib..."
 clang -emit-llvm -S stdio.c
 
 # Remove the previous .ll
-Remove-Item output.ll
+if (Test-Path output.ll) {
+    Remove-Item output.ll
+}
 Rename-Item output.ir output.ll
 
 Write-Host "Compiling output.ir..."
